KG-505. Move Debugger feature in agents-core module
- Move Debugger feature from agents-features module into agents-core module to get access to the feature from the core agents flow and be able to install a feature through an external agent configuration.
