+++ agents/agents-core/src/androidMain/kotlin/ai.koog.agents.core.system/SystemVariablesReader.android.kt
@@ -1,0 +1,21 @@
+package ai.koog.agents.core.system
+
+import io.github.oshai.kotlinlogging.KotlinLogging
+
+@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+internal actual object SystemVariablesReader {
+
+    private val logger = KotlinLogging.logger { }
+
+    internal actual fun getEnvironmentVariable(name: String): String? {
+        val value = java.lang.System.getenv(name)
+        logger.debug { "Getting environment variable '$name' value: '$value'" }
+        return value
+    }
+
+    internal actual fun getVMOption(name: String): String? {
+        val value = java.lang.System.getProperty(name)
+        logger.debug { "Getting VM Option '$name' value: '$value'" }
+        return value
+    }
+}
