+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/FEATURES.md
@@ -17,7 +17,7 @@ This document describes how to use and implement custom features for AIAgent.
     - [Pipeline Interceptors](#pipeline-interceptors)
     - [Advanced Interceptors](#advanced-interceptors)
 - [Available Features](#available-features)
-    - [AgentMemory](#agentmemory)
+    - [Debugger](#debugger)
 
 ## Introduction
 
@@ -532,51 +532,156 @@ class LoggingFeature(val logger: Logger) {
 
 ## Available Features
 
-### AgentMemory
+### Debugger
 
-The AgentMemory provides persistent memory capabilities for agents. It allows agents to store and retrieve information across runs.
+The Debugger feature integrates into an AI agent's pipeline and intercepts various events such as agent start/finish, strategy execution, node execution, LLM calls, and tool operations. These events are collected and can be sent to a remote debugging server for real-time monitoring and analysis.
 
-> **Note**: AgentMemory is in a separate module and requires a separate dependency. It's defined in the `agents-features/agents-features-memory` module.
+Key capabilities of the Debugger feature include:
+- Monitoring the complete lifecycle of AI agent execution
+- Tracking strategy and node executions
+- Recording LLM calls and responses
+- Logging tool operations and their results
+- Capturing errors and exceptions during agent execution
+- Connecting to a remote debugging server for real-time monitoring
 
-Installation:
+### Using in your project
+
+To use the Debugger feature in your project, you need to install it when creating an AI agent. The feature can be configured with custom settings or used with default values.
+
+#### Basic Installation
 
 ```kotlin
-install(AgentMemory) {
-    memoryProvider = LocalFileMemoryProvider(
-        config = LocalMemoryConfig("my-agent-memory"),
-        storage = EncryptedStorage(
-            fs = JVMFileSystemProvider.ReadWrite,
-            encryption = Aes256GCMEncryptor(secretKey)
-        ),
-        fs = JVMFileSystemProvider.ReadWrite,
-        root = Path("path/to/memory/root")
-    )
-
-    featureName = "my-feature"
-    productName = "my-product"
-    organizationName = "my-organization"
+// When creating an agent
+val agent = createAgent(
+    // ... other agent configuration
+) {
+    // Install the Debugger feature with default settings
+    install(Debugger)
 }
 ```
 
-Usage:
+#### Custom Configuration
+
+You can customize the Debugger by specifying a port and connection timeout for the debugging server:
 
 ```kotlin
-// In a node implementation
-context.withMemory {
-    // Save facts to memory
-    saveFactsFromHistory(
-        concept = myConcept,
-        subject = MemorySubject.Project,
-        scope = MemoryScopeType.PRODUCT
-    )
+val agent = createAgent(
+    // ... other agent configuration
+) {
+    install(Debugger) {
+        // Set a specific port for the debugging server
+        setPort(8080)
 
-    // Load facts from memory
-    loadFactsToAgent(
-        concept = myConcept,
-        scopes = listOf(MemoryScopeType.PRODUCT),
-        subjects = listOf(MemorySubject.Project)
-    )
+        // Set a timeout for waiting for the first connection (optional)
+        // If not set, the server will wait indefinitely or use system variables
+        setConnectionWaitingTimeout(5.seconds)
+    }
 }
 ```
 
-For more details on using AgentMemory, see the examples in the `examples` module.
+#### Port Configuration Priority
+
+The Debugger feature determines the port to use in the following order:
+1. Explicitly set port in the configuration (using `setPort()`)
+2. Environment variable `KOOG_DEBUGGER_PORT`
+3. JVM option `-Dkoog.debugger.port=<port>`
+4. Default Koog remote server port (50881)
+
+#### Connection Timeout Configuration Priority
+
+The Debugger feature determines the connection waiting timeout in the following order:
+1. Explicitly set timeout in the configuration (using `setConnectionWaitingTimeout()`)
+2. Environment variable `KOOG_DEBUGGER_WAIT_CONNECTION_MS` (value in milliseconds)
+3. JVM option `-Dkoog.debugger.wait.connection.ms=<milliseconds>`
+4. Default behavior: wait indefinitely for the first connection
+
+#### Testing Debugger Feature
+
+You can test the Debugger feature by creating a client that connects to the debugging server and collects events:
+
+```kotlin
+// Server configuration (agent with Debugger)
+val port = findAvailablePort()
+val agent = createAgent(
+    // ... agent configuration
+) {
+    install(Debugger) {
+        setPort(port)
+    }
+}
+
+// Client configuration
+val clientConfig = DefaultClientConnectionConfig(
+    host = "127.0.0.1", 
+    port = port
+)
+
+// Create a client to collect events
+FeatureMessageRemoteClient(connectionConfig = clientConfig).use { client ->
+    // Collect and verify events
+    // ...
+    
+    // Run the agent
+    agent.run(userPrompt)
+}
+```
+
+### Example of usage
+
+Here's a complete example of using the Debugger feature in a real-world scenario:
+
+```kotlin
+// Create a strategy for the agent
+val strategy = strategy("example-strategy") {
+    val nodeLLMRequest by nodeLLMRequest("llm-request-node")
+    val nodeToolCall by nodeExecuteTool("tool-call-node")
+    val nodeSendToolResult by nodeLLMSendToolResult("send-tool-result-node")
+
+    edge(nodeStart forwardTo nodeLLMRequest)
+    edge(nodeLLMRequest forwardTo nodeToolCall onToolCall { true })
+    edge(nodeLLMRequest forwardTo nodeFinish onAssistantMessage { true })
+    edge(nodeToolCall forwardTo nodeSendToolResult)
+    edge(nodeSendToolResult forwardTo nodeFinish onAssistantMessage { true })
+    edge(nodeSendToolResult forwardTo nodeToolCall onToolCall { true })
+}
+
+// Create a tool registry
+val toolRegistry = ToolRegistry {
+    tool(SearchTool())
+    tool(CalculatorTool())
+}
+
+// Create an agent with the Debugger feature
+val agent = createAgent(
+    agentId = "example-agent",
+    strategy = strategy,
+    promptId = "example-prompt",
+    systemPrompt = "You are a helpful assistant.",
+    toolRegistry = toolRegistry,
+    model = myLLModel
+) {
+    // Install and configure the Debugger feature
+    install(Debugger) {
+        // Use a specific port or let it use the default
+        // setPort(8080)
+    }
+}
+
+// Use the agent
+agent.use { 
+    // Run the agent with a user prompt
+    val result = agent.run("Calculate 25 * 16 and then search for information about the result.")
+    
+    // Process the result
+    println("Agent result: $result")
+}
+```
+
+While the agent is running, the Debugger will collect events such as:
+- Agent start and finish events
+- Strategy execution events
+- Node execution events
+- LLM calls and responses
+- Tool calls and their results
+
+These events can be monitored through a debugging client connected to the specified port.
