+++ agents/agents-core/src/jvmTest/kotlin/ai/koog/agents/core/system/mock/ClientEventsCollector.kt
