--- agents/agents-features/agents-features-debugger/build.gradle.kts
@@ -1,45 +1,0 @@
-import ai.koog.gradle.publish.maven.Publishing.publishToMaven
-
-group = rootProject.group
-version = rootProject.version
-
-plugins {
-    id("ai.kotlin.multiplatform")
-    alias(libs.plugins.kotlin.serialization)
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(project(":agents:agents-core"))
-
-                api(libs.kotlinx.serialization.json)
-
-                api(libs.ktor.client.content.negotiation)
-                api(libs.ktor.serialization.kotlinx.json)
-                api(libs.ktor.server.sse)
-            }
-        }
-
-        commonTest {
-            dependencies {
-                implementation(kotlin("test"))
-                implementation(project(":agents:agents-test"))
-                implementation(project(":prompt:prompt-executor:prompt-executor-clients:prompt-executor-openai-client"))
-                implementation(libs.kotlinx.coroutines.test)
-            }
-        }
-
-        jvmTest {
-            dependencies {
-                implementation(kotlin("test-junit5"))
-                implementation(libs.ktor.client.cio)
-            }
-        }
-    }
-
-    explicitApi()
-}
-
-publishToMaven()
