--- agents/agents-features/agents-features-debugger/src/commonMain/kotlin/ai/koog/agents/features/debugger/feature/writer/DebuggerFeatureMessageRemoteWriter.kt
@@ -1,13 +1,0 @@
-package ai.koog.agents.features.debugger.feature.writer
-
-import ai.koog.agents.core.feature.remote.server.config.ServerConnectionConfig
-import ai.koog.agents.core.feature.writer.FeatureMessageRemoteWriter
-
-/**
- * Handles writing debug feature messages remotely to a server, inheriting common functionality
- * from the `FeatureMessageRemoteWriter` base class.
- *
- * @param connectionConfig The configuration settings for establishing the remote server connection.
- */
-public class DebuggerFeatureMessageRemoteWriter(connectionConfig: ServerConnectionConfig) :
-    FeatureMessageRemoteWriter(connectionConfig)
