+++ build.gradle.kts
@@ -207,7 +207,6 @@ tasks {
 dependencies {
     dokka(project(":agents:agents-core"))
     dokka(project(":agents:agents-ext"))
-    dokka(project(":agents:agents-features:agents-features-debugger"))
     dokka(project(":agents:agents-features:agents-features-event-handler"))
     dokka(project(":agents:agents-features:agents-features-memory"))
     dokka(project(":agents:agents-features:agents-features-opentelemetry"))
