+++ koog-agents/build.gradle.kts
@@ -40,7 +40,6 @@ val excluded = setOf(
 
 val included = setOf(
     ":agents:agents-core",
-    ":agents:agents-features:agents-features-debugger",
     ":agents:agents-features:agents-features-event-handler",
     ":agents:agents-features:agents-features-memory",
     ":agents:agents-features:agents-features-opentelemetry",
